Moved test to webclient module.
