+++ examples/webserver/protocols/src/main/resources/logging.properties
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2022, 2024 Oracle and/or its affiliates.
+# Copyright (c) 2022, 2023 Oracle and/or its affiliates.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,11 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-handlers=io.helidon.logging.jul.HelidonConsoleHandler
-java.util.logging.SimpleFormatter.format=%1$tY.%1$tm.%1$td %1$tH:%1$tM:%1$tS %4$s %3$s !thread!: %5$s%6$s%n
-
+handlers=java.util.logging.ConsoleHandler
+java.util.logging.SimpleFormatter.format=%1$tY.%1$tm.%1$td %1$tH:%1$tM:%1$tS.%1$tL %5$s%6$s%n
 # Global logging level. Can be overridden by specific loggers
 .level=INFO
 io.helidon.webserver.level=INFO
-#io.helidon.webclient.grpc.level=FINEST
-#io.helidon.webserver.grpc.level=FINEST
