+++ webclient/tests/grpc/src/main/proto/events.proto
@@ -1,0 +1,60 @@
+/*
+ * Copyright (c) 2023 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+
+syntax = "proto3";
+option java_package = "io.helidon.webclient.grpc.tests";
+
+import "google/protobuf/empty.proto";
+
+service EventService {
+  rpc Send (Message) returns (google.protobuf.Empty) {}
+  rpc Events (stream EventRequest) returns (stream EventResponse) {}
+}
+
+message Message {
+  string text = 2;
+}
+
+message EventRequest {
+  int64 id = 1;
+  enum Action {
+    SUBSCRIBE = 0;
+    UNSUBSCRIBE = 1;
+  }
+  Action action = 2;
+}
+
+message EventResponse {
+  oneof response_type {
+    Subscribed subscribed = 1;
+    Unsubscribed unsubscribed = 2;
+    Event event = 3;
+  }
+}
+
+message Subscribed {
+  int64 id = 1;
+}
+
+message Unsubscribed {
+  int64 id = 1;
+}
+
+message Event {
+  int64 id = 1;
+  string text = 2;
+}
