+++ webclient/tests/grpc/src/test/java/io/helidon/webclient/grpc/tests/GrpcStubTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.helidon.examples.webserver.protocols;
+package io.helidon.webclient.grpc.tests;
 
 import java.util.ArrayList;
 import java.util.Iterator;
@@ -28,8 +28,8 @@ import java.util.concurrent.TimeoutException;
 import io.grpc.stub.StreamObserver;
 import io.helidon.common.configurable.Resource;
 import io.helidon.common.tls.Tls;
-import io.helidon.examples.grpc.strings.StringServiceGrpc;
-import io.helidon.examples.grpc.strings.Strings;
+import io.helidon.webclient.grpc.tests.StringServiceGrpc;
+import io.helidon.webclient.grpc.tests.Strings;
 import io.helidon.webclient.api.WebClient;
 import io.helidon.webclient.grpc.GrpcClient;
 import io.helidon.webserver.WebServer;
